python/requirements.txt
google-genai>=0.8.0
python-dotenv>=1.0.0
websockets>=12.0
aiohttp>=3.9.0
